+++ src/barr_build_system/barr_build_ctx.h
@@ -29,7 +29,7 @@ typedef struct BARR_CompileJob
 {
     bool dry_run;  // could use bit flags if we need more
     char *src;
-    char out_file[BARR_BUF_SIZE_2048];
+    char out_file[BARR_BUF_SIZE_8192];
     BARR_CompileInfoCTX *ctx;
     BARR_BuildProgressCTX *progress_ctx;
 } BARR_CompileJob;