+++ src/barr_commands/barr_cmd_version.c
@@ -1,13 +1,15 @@
 #include "barr_cmd_version.h"
 #include "barr_io.h"
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 
 barr_i32 BARR_command_version(barr_i32 argc, char **argv)
 {
-    BARR_printf("Build Barrage (barr) version: %d.%d.%d\n", BARR_VERSION_MAJOR, BARR_VERSION_MINOR, BARR_VERSION_PATCH);
-    BARR_printf("Update: %s\n", BARR_VERSION_DATE);
+    BARR_printf("Build Barrage (barr) version: %d.%d.%d-%s\n", BARR_VERSION_MAJOR, BARR_VERSION_MINOR,
+                BARR_VERSION_PATCH, BARR_VERSION_BUILD_TYPE);
+    BARR_printf("Latest update: %s\n", BARR_VERSION_DATE);
 
     BARR_VOID(argc);
     BARR_VOID(argv);