+++ src/olmos/olmos_variables.h
@@ -1,0 +1,23 @@
+#ifndef OLMOS_VARIABLES_H_
+#define OLMOS_VARIABLES_H_
+
+#define OLM_VAR_COMPILER "compiler"
+#define OLM_VAR_LINKER "linker"
+#define OLM_VAR_TARGET "target"
+#define OLM_VAR_TARGET_TYPE "target_type"
+#define OLM_VAR_BUILD_TYPE "build_type"
+#define OLM_VAR_CFLAGS "cflags"
+#define OLM_VAR_CFLAGS_RELEASE "cflags_release"
+#define OLM_VAR_GLOB_SOURCES "glob_sources"
+#define OLM_VAR_SOURCES "sources"
+#define OLM_VAR_INCLUDES "includes"
+#define OLM_VAR_AUTO_INC_DISCOVERY "auto_include_discovery"
+#define OLM_VAR_DEFINES "defines"
+#define OLM_VAR_LIB_PATHS "lib_paths"
+#define OLM_VAR_LINKER_FLAGS "lflags"
+#define OLM_VAR_OUT_DIR "out_dir"
+#define OLM_VAR_BIN_OUT_PATH "binary_out_path"
+#define OLM_VAR_EXCLUDE_PATTERNS "exclude_patterns"
+#define OLM_VAR_VERSION "version"
+
+#endif  // OLMOS_VARIABLES_H_